add payment rpc module
