+++ src/rpc/payment.rs
@@ -1,0 +1,187 @@
+use crate::fiber::{
+    channel::ChannelActorStateStore,
+    graph::PaymentSessionStatus,
+    network::SendPaymentCommand,
+    serde_utils::{U128Hex, U64Hex},
+    types::{Hash256, Pubkey},
+    NetworkActorCommand, NetworkActorMessage,
+};
+use crate::{handle_actor_call, log_and_error};
+use ckb_jsonrpc_types::Script;
+use jsonrpsee::{
+    core::async_trait,
+    proc_macros::rpc,
+    types::{error::CALL_EXECUTION_FAILED_CODE, ErrorObjectOwned},
+};
+use ractor::{call, ActorRef};
+use serde::{Deserialize, Serialize};
+use serde_with::serde_as;
+
+#[serde_as]
+#[derive(Serialize, Deserialize, Debug)]
+pub struct GetPaymentCommandParams {
+    /// The payment hash of the payment to retrieve
+    pub payment_hash: Hash256,
+}
+
+#[serde_as]
+#[derive(Serialize, Deserialize, Clone)]
+pub struct GetPaymentCommandResult {
+    /// The payment hash of the payment
+    pub payment_hash: Hash256,
+    /// The status of the payment
+    pub status: PaymentSessionStatus,
+    #[serde_as(as = "U64Hex")]
+    /// The time the payment was created at, in milliseconds from UNIX epoch
+    created_at: u64,
+    #[serde_as(as = "U64Hex")]
+    /// The time the payment was last updated at, in milliseconds from UNIX epoch
+    pub last_updated_at: u64,
+    /// The error message if the payment failed
+    pub failed_error: Option<String>,
+    /// fee paid for the payment
+    #[serde_as(as = "U128Hex")]
+    pub fee: u128,
+}
+
+#[serde_as]
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub(crate) struct SendPaymentCommandParams {
+    /// the identifier of the payment target
+    target_pubkey: Option<Pubkey>,
+
+    /// the amount of the payment
+    #[serde_as(as = "Option<U128Hex>")]
+    amount: Option<u128>,
+
+    /// the hash to use within the payment's HTLC
+    payment_hash: Option<Hash256>,
+
+    /// the TLC expiry delta should be used to set the timelock for the final hop, in milliseconds
+    #[serde_as(as = "Option<U64Hex>")]
+    final_tlc_expiry_delta: Option<u64>,
+
+    /// the TLC expiry limit for the whole payment, in milliseconds, each hop is with a default tlc delta of 1 day
+    /// suppose the payment router is with N hops, the total tlc expiry limit is at least (N-1) days
+    /// this is also the default value for the payment if this parameter is not provided
+    #[serde_as(as = "Option<U64Hex>")]
+    tlc_expiry_limit: Option<u64>,
+
+    /// the encoded invoice to send to the recipient
+    invoice: Option<String>,
+
+    /// the payment timeout in seconds, if the payment is not completed within this time, it will be cancelled
+    #[serde_as(as = "Option<U64Hex>")]
+    timeout: Option<u64>,
+
+    /// the maximum fee amounts in shannons that the sender is willing to pay
+    #[serde_as(as = "Option<U128Hex>")]
+    max_fee_amount: Option<u128>,
+
+    /// max parts for the payment, only used for multi-part payments
+    #[serde_as(as = "Option<U64Hex>")]
+    max_parts: Option<u64>,
+
+    /// keysend payment
+    keysend: Option<bool>,
+
+    /// udt type script for the payment
+    udt_type_script: Option<Script>,
+
+    /// allow self payment, default is false
+    allow_self_payment: Option<bool>,
+
+    /// dry_run for payment, used for check whether we can build valid router and the fee for this payment,
+    /// it's useful for the sender to double check the payment before sending it to the network,
+    /// default is false
+    dry_run: Option<bool>,
+}
+
+/// RPC module for channel management.
+#[rpc(server)]
+trait PaymentRpc {
+    /// Sends a payment to a peer.
+    #[method(name = "send_payment")]
+    async fn send_payment(
+        &self,
+        params: SendPaymentCommandParams,
+    ) -> Result<GetPaymentCommandResult, ErrorObjectOwned>;
+
+    /// Retrieves a payment.
+    #[method(name = "get_payment")]
+    async fn get_payment(
+        &self,
+        params: GetPaymentCommandParams,
+    ) -> Result<GetPaymentCommandResult, ErrorObjectOwned>;
+}
+
+pub(crate) struct PaymentRpcServerImpl<S> {
+    actor: ActorRef<NetworkActorMessage>,
+    _store: S,
+}
+
+impl<S> PaymentRpcServerImpl<S> {
+    pub(crate) fn new(actor: ActorRef<NetworkActorMessage>, _store: S) -> Self {
+        PaymentRpcServerImpl { actor, _store }
+    }
+}
+
+#[async_trait]
+impl<S> PaymentRpcServer for PaymentRpcServerImpl<S>
+where
+    S: ChannelActorStateStore + Send + Sync + 'static,
+{
+    async fn send_payment(
+        &self,
+        params: SendPaymentCommandParams,
+    ) -> Result<GetPaymentCommandResult, ErrorObjectOwned> {
+        let message = |rpc_reply| -> NetworkActorMessage {
+            NetworkActorMessage::Command(NetworkActorCommand::SendPayment(
+                SendPaymentCommand {
+                    target_pubkey: params.target_pubkey,
+                    amount: params.amount,
+                    payment_hash: params.payment_hash,
+                    final_tlc_expiry_delta: params.final_tlc_expiry_delta,
+                    tlc_expiry_limit: params.tlc_expiry_limit,
+                    invoice: params.invoice.clone(),
+                    timeout: params.timeout,
+                    max_fee_amount: params.max_fee_amount,
+                    max_parts: params.max_parts,
+                    keysend: params.keysend,
+                    udt_type_script: params.udt_type_script.clone().map(|s| s.into()),
+                    allow_self_payment: params.allow_self_payment.unwrap_or(false),
+                    dry_run: params.dry_run.unwrap_or(false),
+                },
+                rpc_reply,
+            ))
+        };
+        handle_actor_call!(self.actor, message, params).map(|response| GetPaymentCommandResult {
+            payment_hash: response.payment_hash,
+            status: response.status,
+            created_at: response.created_at,
+            last_updated_at: response.last_updated_at,
+            failed_error: response.failed_error,
+            fee: response.fee,
+        })
+    }
+
+    async fn get_payment(
+        &self,
+        params: GetPaymentCommandParams,
+    ) -> Result<GetPaymentCommandResult, ErrorObjectOwned> {
+        let message = |rpc_reply| -> NetworkActorMessage {
+            NetworkActorMessage::Command(NetworkActorCommand::GetPayment(
+                params.payment_hash,
+                rpc_reply,
+            ))
+        };
+        handle_actor_call!(self.actor, message, params).map(|response| GetPaymentCommandResult {
+            payment_hash: response.payment_hash,
+            status: response.status,
+            last_updated_at: response.last_updated_at,
+            created_at: response.created_at,
+            failed_error: response.failed_error,
+            fee: response.fee,
+        })
+    }
+}
