+++ tests/nodes/deployer/config.yml
@@ -14,6 +14,7 @@ rpc:
   enabled_modules:
     - cch
     - channel
+    - payment
     - graph
     - info
     - invoice
